+++ example/flutter_example/lib/main.dart
@@ -1,7 +1,6 @@
 import 'dart:io';
 
 import 'package:deepgram_speech_to_text/deepgram_speech_to_text.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_example/utils.dart';
@@ -9,7 +8,7 @@ import 'package:record/record.dart';
 import 'package:universal_file/universal_file.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 
-const fileName = 'jfk.wav';
+const fileName = 'fr.mp3';
 const fileAssetPath = 'assets/$fileName';
 const url = 'https://www2.cs.uic.edu/~i101/SoundFiles/taunt.wav';
 
@@ -22,63 +21,90 @@ void main() async {
 }
 
 // reference : https://developers.deepgram.com/reference/listen-file
-Map<String, dynamic> params = {
+Map<String, dynamic> baseParams = {
   'model': 'nova-2-general',
-  'language': 'fr',
+  'detect_language': true,
   'filler_words': false,
   'punctuation': true,
 };
 
 // make sure to add your API key in the .env file (which is added in the .gitignore and pubspect.yaml's assets)
 final apiKey = dotenv.get("DEEPGRAM_API_KEY");
-Deepgram deepgram = Deepgram(apiKey, baseQueryParams: params);
+Deepgram deepgram = Deepgram(apiKey, baseQueryParams: baseParams);
+
+void checkApiKey() async {
+  print('Checking API key...');
+  final isValid = await deepgram.isApiKeyValid();
+
+  print('API key is valid: $isValid');
+}
 
 void fromFile() async {
+  print('Transcribing from file...');
   final path = await getLocalFilePath('jfk.wav');
   final file = File(path);
 
-  final json = await deepgram.transcribeFromFile(file);
-  print(json);
+  final res = await deepgram.transcribeFromFile(file);
+  print(res.transcript);
 }
 
 void fromUrl() async {
-  final json = await deepgram.transcribeFromUrl(url);
-  print(json);
+  print('Transcribing from url...');
+  final res = await deepgram.transcribeFromUrl(url);
+  print(res.transcript);
 }
 
 void fromBytes() async {
+  print('Transcribing from bytes...');
   final data = await rootBundle.load(fileAssetPath);
   final bytes = data.buffer.asUint8List();
-  final json = await deepgram.transcribeFromBytes(bytes);
-  print(json);
+  final res = await deepgram.transcribeFromBytes(bytes);
+  print(res.transcript);
 }
 
-void fromStream() async {
-  final audioStream = await AudioRecorder().startStream(RecordConfig(
+final mic = AudioRecorder();
+void startStream() async {
+  await mic.hasPermission();
+
+  final audioStream = await mic.startStream(RecordConfig(
     encoder: AudioEncoder.pcm16bits,
     sampleRate: 16000,
     numChannels: 1,
   ));
-  final jsonStream = deepgram.transcribeFromLiveAudioStream(audioStream);
-  jsonStream.listen((json) {
-    print(json);
+
+  print('Recording started...');
+
+  final liveParams = {
+    'detect_language': false, // not supported by streaming API
+    'language': 'en',
+    // must specify encoding and sample_rate according to the audio stream
+    'encoding': 'linear16',
+    'sample_rate': 16000,
+  };
+
+  final stream = deepgram.transcribeFromLiveAudioStream(audioStream,
+      queryParams: liveParams);
+
+  stream.listen((res) {
+    print(res.transcript);
   });
-}
 
-void fromTranscriber() async {
-  final audioStream = await AudioRecorder().startStream(RecordConfig(
-    encoder: AudioEncoder.pcm16bits,
-    sampleRate: 16000,
-    numChannels: 1,
-  ));
-  final transcriber = deepgram.createLiveTranscriber(audioStream);
+  // alternativly you can use the DeepgramLiveTranscriber class :
+  /*
+  final transcriber = deepgram.createLiveTranscriber(audioStream, queryParams: params);
 
   transcriber.start();
 
-  transcriber.jsonStream.listen((json) {
-    print(json);
+  transcriber.stream.listen((res) {
+    print(res.transcript);
   });
   transcriber.close();
+  */
+}
+
+void stopStream() async {
+  print('Recording stopped');
+  await mic.stop();
 }
 
 class MainApp extends StatelessWidget {
@@ -98,16 +124,25 @@ class MainApp extends StatelessWidget {
           backgroundColor: Colors.blue,
         ),
         body: Center(
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.center,
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: [
-              ElevatedButton(onPressed: fromFile, child: Text('TranscribeFromFile')),
-              ElevatedButton(onPressed: fromUrl, child: Text('TranscribeFromUrl')),
-              ElevatedButton(onPressed: fromBytes, child: Text('TranscribeFromBytes')),
-              ElevatedButton(onPressed: fromStream, child: Text('TranscribeFromLiveAudioStream')),
-              ElevatedButton(onPressed: fromTranscriber, child: Text('CreateLiveTranscriber')),
-            ],
+          child: Padding(
+            padding: const EdgeInsets.all(40.0),
+            child: Column(
+              crossAxisAlignment: CrossAxisAlignment.center,
+              mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+              children: [
+                ElevatedButton(
+                    onPressed: checkApiKey, child: Text('Check Api Key')),
+                Divider(),
+                ElevatedButton(onPressed: fromFile, child: Text('From File')),
+                ElevatedButton(onPressed: fromUrl, child: Text('From Url')),
+                ElevatedButton(onPressed: fromBytes, child: Text('From Bytes')),
+                Divider(),
+                ElevatedButton(
+                    onPressed: startStream, child: Text('Start Stream')),
+                ElevatedButton(
+                    onPressed: stopStream, child: Text('Stop Stream')),
+              ],
+            ),
           ),
         ),
       ),